--- a/aten/src/ATen/Declarations.cwrap
+++ b/aten/src/ATen/Declarations.cwrap
@@ -570,7 +570,6 @@
   types:
     - floating_point
   backends:
-    - CPU
     - CUDA
   variants:
     - function