--- a/aten/src/ATen/native/ReduceOps.cpp
+++ b/aten/src/ATen/native/ReduceOps.cpp
@@ -976,18 +976,7 @@ std::tuple<Tensor&, Tensor&> cummin_out(Tensor& values, Tensor& indices, const T
 }
 
 Tensor dist(const Tensor &self, const Tensor& other, Scalar p){
-
-  if (at::isfinite(self).logical_not().any().item<uint8_t>() || at::isfinite(other).logical_not().any().item<uint8_t>()){
-    Tensor result = at::empty({1}, self.options());
-    result[0] = std::numeric_limits<double>::infinity();
-    return result[0];
-  }
-
-  Tensor result = at::empty_like(self, at::MemoryFormat::Contiguous);
-  result = at::abs(self - other);
-  result = at::pow_out(result, result, p);
-  result = at::pow_out(result, at::sum(result), 1.0 / p.toDouble());
-  return result;
+  return at::norm(self - other, p);
 }
 
 }} // namespace at::native