@kostmo 2020-01-06
## :pill: CircleCI build failures summary and remediations
As of commit a0c4a67c:


* **2/2** failures introduced in this PR


## Detailed failure analysis

One may explore the probable reasons each build failed interactively [on the Dr. CI website](https://dr.pytorch.org/commit-details.html?sha1=a0c4a67cd0dddb282169ae0652cd5e3764b7db06).

### :detective: 2 new failures recognized by patterns
The following build failures do not appear to be due to upstream breakage:
#### [![See CircleCI build](https://avatars0.githubusercontent.com/ml/7?s=12)](https://circleci.com/gh/pytorch/pytorch/4140324) pytorch_linux_xenial_py2_7_9_test (1/2)
**Step:** "Test" ([full log](https://dr.pytorch.org/api/view-log-full?build_id=80741729) | [pattern match details](https://dr.pytorch.org/build-details.html?build_id=80741729))
<details>
<summary>
<code>Jan 06 15:04:45 RuntimeError: test_torch failed!</code>
</summary>

```
Jan 06 15:04:45 Ran 1019 tests in 78.119s 
Jan 06 15:04:45  
Jan 06 15:04:45 FAILED (failures=1, skipped=46) 
Jan 06 15:04:45  
Jan 06 15:04:45 Generating XML reports... 
Jan 06 15:04:45 Traceback (most recent call last): 
Jan 06 15:04:45   File "test/run_test.py", line 456, in <module> 
Jan 06 15:04:45     main() 
Jan 06 15:04:45   File "test/run_test.py", line 449, in main 
Jan 06 15:04:45     raise RuntimeError(message) 
Jan 06 15:04:45 RuntimeError: test_torch failed! 
Jan 06 15:04:46 + cleanup 
Jan 06 15:04:46 + retcode=1 
Jan 06 15:04:46 + set +x 
Jan 06 15:04:46 =================== sccache compilation log =================== 
Jan 06 15:04:46 ERROR:sccache::server: Compilation failed: Output { status: ExitStatus(ExitStatus(256)), stdout: "", stderr: "/tmp/torch_extensions/test_compilation_error_formatting/main.cpp: In function \'int main()\':\n/tmp/torch_extensions/test_compilation_error_formatting/main.cpp:2:23: error: expected \';\' before \'}\' token\n int main() { return 0 }\n                       ^\n" } 
Jan 06 15:04:46  
Jan 06 15:04:46 =========== If your build fails, please take a look at the log above for possible reasons =========== 
Jan 06 15:04:46 Compile requests                 51 
Jan 06 15:04:46 Compile requests executed        27 
Jan 06 15:04:46 Cache hits                        3 
```

</details>

#### [![See CircleCI build](https://avatars0.githubusercontent.com/ml/7?s=12)](https://circleci.com/gh/pytorch/pytorch/4140321) pytorch_linux_xenial_py3_6_gcc5_4_test (2/2)
**Step:** "Test" ([full log](https://dr.pytorch.org/api/view-log-full?build_id=80741806) | [pattern match details](https://dr.pytorch.org/build-details.html?build_id=80741806))
<details>
<summary>
<code>Jan 06 15:05:04 RuntimeError: test_torch failed!</code>
</summary>

```
Jan 06 15:05:04 Ran 1019 tests in 74.842s 
Jan 06 15:05:04  
Jan 06 15:05:04 FAILED (failures=1, skipped=36) 
Jan 06 15:05:04  
Jan 06 15:05:04 Generating XML reports... 
Jan 06 15:05:04 Traceback (most recent call last): 
Jan 06 15:05:04   File "test/run_test.py", line 456, in <module> 
Jan 06 15:05:04     main() 
Jan 06 15:05:04   File "test/run_test.py", line 449, in main 
Jan 06 15:05:04     raise RuntimeError(message) 
Jan 06 15:05:04 RuntimeError: test_torch failed! 
Jan 06 15:05:04 + cleanup 
Jan 06 15:05:04 + retcode=1 
Jan 06 15:05:04 + set +x 
Jan 06 15:05:04 =================== sccache compilation log =================== 
Jan 06 15:05:04 ERROR:sccache::server: Compilation failed: Output { status: ExitStatus(ExitStatus(256)), stdout: "", stderr: "/tmp/torch_extensions/test_compilation_error_formatting/main.cpp: In function \'int main()\':\n/tmp/torch_extensions/test_compilation_error_formatting/main.cpp:2:23: error: expected \';\' before \'}\' token\n int main() { return 0 }\n                       ^\n" } 
Jan 06 15:05:04  
Jan 06 15:05:04 =========== If your build fails, please take a look at the log above for possible reasons =========== 
Jan 06 15:05:04 Compile requests                 48 
Jan 06 15:05:04 Compile requests executed        25 
Jan 06 15:05:04 Cache hits                        1 
```

</details>



---

<details><summary>This comment was automatically generated by <a href="https://github.com/kostmo/circleci-failure-tracker/tree/master/docs/from-pull-request-comment">Dr. CI</a> (expand for details).</summary>Follow <a href="https://dr.pytorch.org/admin/comments-opt-out.html">this link to opt-out</a> of these comments for your Pull Requests.

Please report bugs/suggestions on the <a href="https://github.com/kostmo/circleci-failure-tracker/issues">GitHub issue tracker</a>.
</details>