@albanD 2020-01-23
You will need to rebase as there are conflicts with master.
Also the CI failure look relevant. Can you take a look?