@rivergold 2020-01-30
> You will need to rebase as there are conflicts with master.
> Also the CI failure look relevant. Can you take a look?

I have rebased and updated the code, but there is still a failure. Can give some guidances？Thanks a lot.