@albanD 2020-01-30
You can ignore the macos build failure.