@albanD 2020-02-03
Looks good.
For precision I meant in terms of value. Because this is floating point numbers, the result is not exact. And for large Tensor, we don't want the result to be off by too much. In particular, the TH implementation for `float32`, the TH implementation accumulates in a `double` to reduce precision loss. I was wondering of `norm()` does that too or not.