@VitalyFedyunin 2020-02-03
Looks good to me. BTW you can put CPU and CUDA implementations together (at least it seems so).