@albanD 2020-02-04
I was thinking of running something like
```python
import torch

torch.manual_seed(123)
a = torch.rand(100, 100, 100)
b = torch.rand(100, 100, 100)

print(torch.dist(a, b))
print(torch.dist(a.double(), b.double()))
```
With the previous and new code. To see if the different is much bigger with the new code or not.

As @VitalyFedyunin mentioned, even though it returns an `accreal` which is higher precision, it still accumulates in regular precision `scalar_t sum;`. So you should see similar results :)