@rivergold 2020-02-05
@albanD  @VitalyFedyunin 
CPU and Cuda implementations are updated together.