@albanD 2020-02-05
Thanks. Could you rebase on top of master to make sure all test pass please?

Also did you had a chance to run the precision test above with this PR?